feat(admin-stats): merge campaign and platform stats into unified dashboard (#3571)
