+++ packages/app/src/modules/admin/AdminNavigation.tsx
@@ -8,8 +8,7 @@ const adminLinks = [
 	{ href: "/admin/declarations", label: "Déclarations" },
 	{ href: "/admin/impersonate", label: "Mimoquer un Siren" },
 	{ href: "/admin/liste-referents", label: "Référents" },
-	{ href: "/admin/stats/campagne", label: "Statistiques campagne" },
-	{ href: "/admin/stats/plateforme", label: "Statistiques plateforme" },
+	{ href: "/admin/stats", label: "Statistiques" },
 	{ href: "/admin/parametres", label: "Paramètres" },
 ] as const;
