+++ packages/app/src/modules/admin/stats/index.ts
@@ -3,12 +3,11 @@ export { AdminKpiTile } from "./AdminKpiTile";
 export { CampaignProgressionChart } from "./CampaignProgressionChart";
 export { CampaignProgressionTable } from "./CampaignProgressionTable";
 export { CampaignRateTile } from "./CampaignRateTile";
-export { CampaignStatsPage } from "./CampaignStatsPage";
 export { CompletionFunnelChart } from "./CompletionFunnelChart";
 export { CompletionFunnelTable } from "./CompletionFunnelTable";
 export { formatCount, formatDays, formatPercent } from "./formatters";
-export { PlatformStatsPage } from "./PlatformStatsPage";
 export { StagnationDaysFilter } from "./StagnationDaysFilter";
+export { StatsDashboard } from "./StatsDashboard";
 export { StepDropoffChart } from "./StepDropoffChart";
 export { StepDropoffTable } from "./StepDropoffTable";
 export { StepDurationsChart } from "./StepDurationsChart";
